fix long -> int

+++ library/src/main/java/cn/com/mark/multiimage/core/PreferencesUtils.java
@@ -63,7 +63,7 @@ public class PreferencesUtils {
 		editor.commit();
 	}
 
-	public static long getInt(String key, int defValue) {
+	public static int getInt(String key, int defValue) {
 		return instance.sharedPreferences.getInt(key, defValue);
 	}
